changed floats to doubles, fixed Navigation solution

+++ controller/CoreRobotics/MovementModeler.cs
@@ -26,9 +26,9 @@ namespace Robocup.CoreRobotics
             //this just has it move randomly:
 
             Vector2 newposition = info.Position + dt * info.Velocity;
-            Vector2 newvelocity = new Vector2((float)(r.NextDouble() * 4 - 2), (float)(r.NextDouble() * 4 - 2));
+            Vector2 newvelocity = new Vector2((double)(r.NextDouble() * 4 - 2), (double)(r.NextDouble() * 4 - 2));
             double newrotvelocity = 20 * r.NextDouble() - 10;
-            return new RobotInfo(newposition, newvelocity, newrotvelocity, info.Orientation + (float)(info.RotationalVelocity * dt), info.ID);
+            return new RobotInfo(newposition, newvelocity, newrotvelocity, info.Orientation + (double)(info.RotationalVelocity * dt), info.ID);
             
 
             //idea for implementation:
