+++ strategy/Geometry/UsefulFunctions.cs
@@ -3,25 +3,31 @@ using System.Collections.Generic;
 using System.Text;
 using Robocup.Core;
 
-namespace Robocup.Geometry {
-    public static class UsefulFunctions {
+namespace Robocup.Geometry
+{
+    public static class UsefulFunctions
+    {
 
-        static public float distance(Vector2 p1, Vector2 p2) {
-            return (float)Math.Sqrt((p1.X - p2.X) * (p1.X - p2.X) + (p1.Y - p2.Y) * (p1.Y - p2.Y));
+        static public double distance(Vector2 p1, Vector2 p2)
+        {
+            return Math.Sqrt((p1.X - p2.X) * (p1.X - p2.X) + (p1.Y - p2.Y) * (p1.Y - p2.Y));
         }
-        static public float distancesq(Vector2 p1, Vector2 p2) {
+        static public double distancesq(Vector2 p1, Vector2 p2)
+        {
             return (p1.X - p2.X) * (p1.X - p2.X) + (p1.Y - p2.Y) * (p1.Y - p2.Y);
         }
         /// <summary>
         /// Returns the z-component of the crossproduct P2P1 x P2P3
         /// </summary>
-        static public float crossproduct(Vector2 p1, Vector2 p2, Vector2 p3) {
+        static public double crossproduct(Vector2 p1, Vector2 p2, Vector2 p3)
+        {
             return (p1.X - p2.X) * (p3.Y - p2.Y) - (p3.X - p2.X) * (p1.Y - p2.Y);
         }
         /// <summary>
         /// Returns the dotproduct of P2P1 and P2P3
         /// </summary>
-        static public float dotproduct(Vector2 p1, Vector2 p2, Vector2 p3) {
+        static public double dotproduct(Vector2 p1, Vector2 p2, Vector2 p3)
+        {
             return (p1.X - p2.X) * (p3.X - p2.X) + (p1.Y - p2.Y) * (p3.Y - p2.Y);
         }
 
@@ -31,7 +37,8 @@ namespace Robocup.Geometry {
         /// Uses
         /// Returns a value in the range [-Pi,Pi)
         /// </summary>
-        static public double angleDifference(double angle1, double angle2) {
+        static public double angleDifference(double angle1, double angle2)
+        {
             //first get the inputs in the range [0, 2Pi):
             while (angle1 < 0)
                 angle1 += Math.PI * 2;
