+++ strategy/Navigation/Obstacle.cs
@@ -3,7 +3,8 @@ using System.Collections.Generic;
 using System.Text;
 using Robocup.Core;
 
-namespace Navigation {
+namespace Navigation
+{
     /// <summary>
     /// A simple class to represent obstacles, it gives a useful way to
     /// treat robots and the ball as similar things.  Also, setting the
@@ -11,12 +12,14 @@ namespace Navigation {
     /// obstacle.
     /// </summary>
     //yes it should probably be a struct, but then you can't return null when you ask for the blocking obstacle
-    public class Obstacle {
-        public Obstacle(Vector2 pos, float size) {
+    public class Obstacle
+    {
+        public Obstacle(Vector2 pos, double size)
+        {
             this.position = pos;
             this.size = size;
         }
         public Vector2 position;
-        public float size;
+        public double size;
     }
 }
